+++ inst/shiny/modules/tab_nca/setup/summary.R
@@ -24,7 +24,7 @@ summary_server <- function(id, processed_pknca_data) {
       dose_group_columns <- group_vars(processed_pknca_data()$dose)
 
       data <- processed_pknca_data()$intervals %>%
-        apply_labels(type = "ADPC") %>%
+        apply_labels(type = "ADNCA") %>%
         select(where(~!is.logical(.) | any(. == TRUE))) %>%
         arrange(!!!syms(c(conc_group_columns, "type_interval", "start", "end")))
     })