+++ NEWS.md
@@ -9,6 +9,9 @@
 
 ## Bugs fixed
 
+* Exact duplicate records (DTYPE = "COPY") and time duplicates (DTYPE = "TIME DUPLICATE") are not removed from the data, but just excluded from the analysis (#765) 
+
+
 # aNCA 0.1.0
 
 * Initial CRAN submission.