+++ NEWS.md
@@ -10,7 +10,9 @@
 ## Bugs fixed
 * Bug fix for box/violin plots that were crashing when PPSTRES is NA (#785)
 
-* Exact duplicate records (DTYPE = "COPY") and time duplicates (DTYPE = "TIME DUPLICATE") are not removed from the data, but just excluded from the analysis (#765) 
+* Exact duplicate records (DTYPE = "COPY") and time duplicates (DTYPE = "TIME DUPLICATE") are not removed from the data, but just excluded from the analysis (#765)
+
+* ZIP folder with results will include now the exploration tab outputs: individual plots, mean plots (#794)
 
 
 # aNCA 0.1.0